refactor(chunk_graph): split methods of  into different files
